+++ src/main/java/com/greatminds/ayni/ayniedgeapi/iot/domain/model/commands/CreateStatusCommand.java
@@ -1,4 +1,4 @@
 package com.greatminds.ayni.ayniedgeapi.iot.domain.model.commands;
 
-public record CreateStatusCommand(Long cropId, double temperature, double humidity, double oxygen, double waterLevel) {
+public record CreateStatusCommand(Long cropId, Float temperature, Float humidity, Float oxygen, Float waterLevel) {
 }
